Use CGPostKeyboardEvent instead of CGEventPost
For certain events (e.g. arrow keys) the modifier mask that is set on
the event (which includes the function key) does not seem to get cleared
on key up, which ruins all future events because their mask contains
those flags even though they shouldn't be active anymore. The deprecated
function doesn't seem to do this so use it for now until I can figure
out what is going on.

Fixes #11. (Or, at least, makes it significantly less annoying.)